Promotes builder-spec classes to their own file.
