+++ automotive/maliput/multilane/BUILD.bazel
@@ -60,10 +60,12 @@ drake_cc_library(
     name = "builder",
     srcs = [
         "builder.cc",
+        "builder_spec.cc",
         "connection.cc",
     ],
     hdrs = [
         "builder.h",
+        "builder_spec.h",
         "connection.h",
     ],
     deps = [
@@ -109,6 +111,16 @@ drake_cc_googletest(
         "//automotive/maliput/api/test_utilities",
         "//automotive/maliput/multilane/test_utilities",
         "//common/test_utilities:eigen_matrix_compare",
+    ],
+)
+
+drake_cc_googletest(
+    name = "multilane_builder_spec_test",
+    srcs = ["test/multilane_builder_spec_test.cc"],
+    deps = [
+        ":builder",
+        "//automotive/maliput/api/test_utilities",
+        "//automotive/maliput/multilane/test_utilities",
     ],
 )
